refactor(processor): remove old voice-activity detector after Phase 1 unified VAD
  Delete the old two-part detector.

  - Delete old speech run-formation code (`findSpeechCandidatesFromIntervals`,
    speech scoring, `roomToneSpan` type, dead run-builder constants)
  - Delete old room-tone detector files entirely
    (`analyser_candidates_room_tone_scoring.go`,
    `analyser_candidates_room_tone.go`)
  - Move retained seed estimators (`estimateNoiseFloorAndThreshold`,
    `computeSilenceMedians`) to new `analyser_noise_seed.go` for reuse by
    buildInputMeasurements
  - Remove tests that exercised only the old run-builder; fix stale test
    fixtures (centroid bounds changed in prior commit)

  Unified VAD detector is now the sole live path. Build, lint, and tests
  all green.
