+++ internal/processor/analyser.go
@@ -367,88 +367,6 @@ func AnalyseAudio(ctx stdcontext.Context, filename string, config *BaseFilterCon
 	return measurements, nil
 }
 
-type noiseProfileSelection struct {
-	roomToneResult *findBestRoomToneRegionResult
-	noiseProfile   *NoiseProfile
-}
-
-func selectNoiseProfile(measurements *AudioMeasurements, intervals, silenceIntervals []IntervalSample, silMedians silenceMedians, log debugLogger) noiseProfileSelection {
-	measurements.Regions.RoomToneRegions = findRoomToneCandidatesFromIntervals(silenceIntervals, measurements.Noise.RoomToneDetectLevel, silMedians)
-
-	roomToneResult := findBestRoomToneRegion(measurements.Regions.RoomToneRegions, silenceIntervals, log)
-	measurements.Regions.RoomToneCandidates = roomToneResult.Candidates
-	measurements.Noise.VoiceActivated = detectVoiceActivated(roomToneResult.Candidates)
-
-	selection := noiseProfileSelection{roomToneResult: roomToneResult}
-	if roomToneResult.BestRegion == nil {
-		return selection
-	}
-
-	originalRegion := roomToneResult.BestRegion
-	refinedRegion := refineToGoldenSubregion(originalRegion, intervals)
-	wasRefined := refinedRegion.Start != originalRegion.Start || refinedRegion.Duration != originalRegion.Duration
-
-	profile := extractNoiseProfileFromIntervals(refinedRegion, silenceIntervals)
-	if profile == nil {
-		return selection
-	}
-
-	selection.noiseProfile = profile
-	measurements.Regions.NoiseProfile = profile
-
-	// Capture the elected candidate's already-computed RegionSample for the run
-	// record (regions.room_tone.samples.input). The elected candidate is the one in
-	// Candidates whose Region matches the elected BestRegion (the same region the
-	// NoiseProfile is built from). Pure read of an existing measurement; no scoring,
-	// election, or DSP value changes - audio stays bit-exact.
-	for i := range roomToneResult.Candidates {
-		c := &roomToneResult.Candidates[i]
-		if c.Region.Start == originalRegion.Start && c.Region.Duration == originalRegion.Duration {
-			measurements.Regions.ElectedRoomToneSample = &c.RegionSample
-			break
-		}
-	}
-
-	if wasRefined {
-		profile.WasRefined = true
-		profile.OriginalStart = originalRegion.Start
-		profile.OriginalDuration = originalRegion.Duration
-	}
-
-	if profile.MeasuredNoiseFloor != 0 && !math.IsInf(profile.MeasuredNoiseFloor, -1) {
-		measurements.Noise.Floor = profile.MeasuredNoiseFloor
-		measurements.Noise.FloorSource = "silence_profile"
-	}
-
-	return selection
-}
-
-func selectSpeechProfile(measurements *AudioMeasurements, intervals []IntervalSample, noiseSelection noiseProfileSelection, log debugLogger) {
-	// Speech searches the whole file. The elected room-tone region (when present)
-	// only excludes its own intervals from speech runs; it never sets a search floor.
-	var span roomToneSpan
-	if noiseSelection.roomToneResult != nil && noiseSelection.roomToneResult.BestRegion != nil {
-		best := noiseSelection.roomToneResult.BestRegion
-		span = roomToneSpan{Start: best.Start, End: best.End, Present: true}
-	}
-
-	measurements.Regions.SpeechRegions = findSpeechCandidatesFromIntervals(intervals, span, measurements.Noise.VoiceActivated, measurements.Dynamics.RMSLevel, measurements.Noise.Floor)
-
-	speechResult := findBestSpeechRegion(measurements.Regions.SpeechRegions, intervals, noiseSelection.noiseProfile, log)
-	measurements.Regions.SpeechCandidates = speechResult.Candidates
-
-	if speechResult.BestRegion == nil {
-		return
-	}
-
-	for i := range speechResult.Candidates {
-		if speechResult.Candidates[i].Region.Start == speechResult.BestRegion.Start {
-			measurements.Regions.SpeechProfile = &speechResult.Candidates[i]
-			return
-		}
-	}
-}
-
 func buildInputMeasurements(filename string, collection *analysisFrameCollection, config *BaseFilterConfig, defaultNoiseFloor float64) (*AudioMeasurements, error) {
 	acc := collection.accumulators
