+++ internal/processor/analyser_test.go
@@ -3,7 +3,6 @@ package processor
 import (
 	"context"
 	"encoding/json"
-	"math"
 	"testing"
 	"time"
 
@@ -355,162 +354,6 @@ func makeTestIntervals(startTime time.Duration, rmsLevels []float64) []IntervalS
 	return intervals
 }
 
-func TestRefineToGoldenSubregion(t *testing.T) {
-	tests := []struct {
-		name          string
-		candidate     *RoomToneRegion
-		intervals     []IntervalSample
-		wantStart     time.Duration
-		wantDuration  time.Duration
-		wantUnchanged bool // If true, expect original region returned
-		desc          string
-	}{
-		{
-			name: "short candidate - no refinement needed",
-			candidate: &RoomToneRegion{
-				Start:    24 * time.Second,
-				End:      34 * time.Second,
-				Duration: 10 * time.Second,
-			},
-			intervals:     makeTestIntervals(24*time.Second, make([]float64, 40)), // 40 intervals = 10s
-			wantStart:     24 * time.Second,
-			wantDuration:  10 * time.Second,
-			wantUnchanged: true,
-			desc:          "candidates at or below 10s should pass through unchanged",
-		},
-		{
-			name: "long candidate with uniform quality",
-			candidate: &RoomToneRegion{
-				Start:    24 * time.Second,
-				End:      44 * time.Second,
-				Duration: 20 * time.Second,
-			},
-			intervals: makeTestIntervals(24*time.Second, func() []float64 {
-				// 80 intervals at -70 dBFS (uniform)
-				levels := make([]float64, 80)
-				for i := range levels {
-					levels[i] = -70.0
-				}
-				return levels
-			}()),
-			wantStart:    24 * time.Second, // First window when all equal
-			wantDuration: 10 * time.Second,
-			desc:         "uniform quality should return first 10s window",
-		},
-		{
-			name: "long candidate with golden pocket at end",
-			candidate: &RoomToneRegion{
-				Start:    24 * time.Second,
-				End:      44 * time.Second,
-				Duration: 20 * time.Second,
-			},
-			intervals: makeTestIntervals(24*time.Second, func() []float64 {
-				// 80 intervals: first 40 at -65 dBFS, last 40 at -75 dBFS
-				levels := make([]float64, 80)
-				for i := range levels {
-					if i < 40 {
-						levels[i] = -65.0 // Noisier first half
-					} else {
-						levels[i] = -75.0 // Quieter second half
-					}
-				}
-				return levels
-			}()),
-			wantStart:    34 * time.Second, // Should find quieter region at 34s (10s into candidate)
-			wantDuration: 10 * time.Second,
-			desc:         "should find the quieter 10s window in the second half",
-		},
-		{
-			name: "candidate at recording start",
-			candidate: &RoomToneRegion{
-				Start:    0,
-				End:      15 * time.Second,
-				Duration: 15 * time.Second,
-			},
-			intervals: makeTestIntervals(0, func() []float64 {
-				// 60 intervals: first 20 at -60 dBFS, rest at -72 dBFS
-				levels := make([]float64, 60)
-				for i := range levels {
-					if i < 20 {
-						levels[i] = -60.0 // Intro noise
-					} else {
-						levels[i] = -72.0 // Clean room tone
-					}
-				}
-				return levels
-			}()),
-			wantStart:    5 * time.Second, // 20 intervals = 5s offset to quieter region
-			wantDuration: 10 * time.Second,
-			desc:         "should handle candidates starting at 0s correctly",
-		},
-		{
-			name: "insufficient intervals - returns original",
-			candidate: &RoomToneRegion{
-				Start:    24 * time.Second,
-				End:      30 * time.Second,
-				Duration: 6 * time.Second,
-			},
-			intervals:     makeTestIntervals(24*time.Second, make([]float64, 24)), // 24 intervals = 6s < 8s minimum
-			wantStart:     24 * time.Second,
-			wantDuration:  6 * time.Second,
-			wantUnchanged: true,
-			desc:          "candidates with insufficient intervals should pass through",
-		},
-		{
-			name:          "nil candidate",
-			candidate:     nil,
-			intervals:     makeTestIntervals(0, make([]float64, 80)),
-			wantUnchanged: true,
-			desc:          "nil input should return nil",
-		},
-		{
-			name: "no intervals in range",
-			candidate: &RoomToneRegion{
-				Start:    100 * time.Second,
-				End:      120 * time.Second,
-				Duration: 20 * time.Second,
-			},
-			intervals:     makeTestIntervals(0, make([]float64, 80)), // Intervals at 0-20s, candidate at 100-120s
-			wantStart:     100 * time.Second,
-			wantDuration:  20 * time.Second,
-			wantUnchanged: true,
-			desc:          "should return original when no intervals match candidate range",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := refineToGoldenSubregion(tt.candidate, tt.intervals)
-
-			if tt.candidate == nil {
-				if result != nil {
-					t.Errorf("expected nil result for nil candidate, got %+v", result)
-				}
-				return
-			}
-
-			if result == nil {
-				t.Fatal("unexpected nil result")
-			}
-
-			if tt.wantUnchanged {
-				if result.Start != tt.candidate.Start || result.Duration != tt.candidate.Duration {
-					t.Errorf("expected unchanged region, got Start=%v Duration=%v (original Start=%v Duration=%v)",
-						result.Start, result.Duration, tt.candidate.Start, tt.candidate.Duration)
-				}
-				return
-			}
-
-			if result.Start != tt.wantStart {
-				t.Errorf("Start = %v, want %v [%s]", result.Start, tt.wantStart, tt.desc)
-			}
-			if result.Duration != tt.wantDuration {
-				t.Errorf("Duration = %v, want %v [%s]", result.Duration, tt.wantDuration, tt.desc)
-			}
-		})
-	}
-}
-
 func TestGetIntervalsInRange(t *testing.T) {
 	// Create intervals from 0s to 20s (80 intervals × 250ms)
 	allIntervals := makeTestIntervals(0, make([]float64, 80))
@@ -632,353 +475,6 @@ func TestScoreIntervalWindow(t *testing.T) {
 }
 
 // ============================================================================
-// Two-Pass Room Tone Candidate Selection Tests
-// ============================================================================
-
-// makeRoomToneTestIntervals creates interval samples for testing findBestRoomToneRegion.
-// All intervals share the same spectral/amplitude values for predictable scoring.
-func makeRoomToneTestIntervals(start time.Duration, duration time.Duration, rms, peak, centroid, flatness, kurtosis, flux float64) []IntervalSample {
-	count := int(duration / (250 * time.Millisecond))
-	intervals := make([]IntervalSample, count)
-	for i := range intervals {
-		intervals[i] = IntervalSample{
-			Timestamp: start + time.Duration(i)*250*time.Millisecond,
-			RMSLevel:  rms,
-			PeakLevel: peak,
-			Spectral: SpectralMetrics{
-				Centroid: centroid,
-				Flatness: flatness,
-				Kurtosis: kurtosis,
-				Flux:     flux,
-			},
-		}
-	}
-	return intervals
-}
-
-func TestFindBestRoomToneRegion_HighestScoreWinsAfterDip(t *testing.T) {
-	// Three candidates with a scoring dip in the middle (B). Under a single-pass
-	// algorithm the dip from A to B would trigger early termination, electing A.
-	// The two-pass algorithm scores all candidates first and then selects the
-	// highest, so C should win.
-	//
-	// Amplitude now rewards proximity to the candidate-population RMS median
-	// (clusterP50). With RMS {A:-78, B:-50, C:-66} the median is C's -66, so C
-	// sits exactly on the cluster and earns the top amplitude term; A is a quiet
-	// outlier (12 dB below the cluster) and B is the loud dip.
-
-	regions := []RoomToneRegion{
-		{Start: 0, End: 10 * time.Second, Duration: 10 * time.Second},
-		{Start: 15 * time.Second, End: 25 * time.Second, Duration: 10 * time.Second},
-		{Start: 30 * time.Second, End: 40 * time.Second, Duration: 10 * time.Second},
-	}
-
-	// Candidate A intervals: quiet outlier, far below the cluster median
-	intervalsA := makeRoomToneTestIntervals(0, 10*time.Second,
-		-78.0, -73.0, 100.0, 0.9, 1.0, 0.0)
-	// Candidate B intervals: the dip (loud, voice-range centroid, poor spectral)
-	intervalsB := makeRoomToneTestIntervals(15*time.Second, 10*time.Second,
-		-50.0, -45.0, 2375.0, 0.3, 5.0, 0.005)
-	// Candidate C intervals: on the cluster median, strong spectral - highest
-	intervalsC := makeRoomToneTestIntervals(30*time.Second, 10*time.Second,
-		-66.0, -61.0, 100.0, 0.9, 1.0, 0.0)
-
-	allIntervals := append(append(intervalsA, intervalsB...), intervalsC...)
-
-	result := findBestRoomToneRegion(regions, allIntervals, nil)
-
-	if result.BestRegion == nil {
-		t.Fatal("expected a best region to be selected")
-	}
-	if result.BestRegion.Start != 30*time.Second {
-		t.Errorf("BestRegion.Start = %v, want 30s (highest-scoring candidate C)", result.BestRegion.Start)
-	}
-	if len(result.Candidates) != 3 {
-		t.Errorf("len(Candidates) = %d, want 3 (all three scored)", len(result.Candidates))
-	}
-
-	// Log scores for debugging
-	for _, c := range result.Candidates {
-		t.Logf("candidate start=%v score=%.4f", c.Region.Start, c.Score)
-	}
-}
-
-func TestFindBestRoomToneRegion_EarlierCandidatePreferredWithinTolerance(t *testing.T) {
-	// Two candidates with scores differing by < 0.02 (within selectionTolerance).
-	// The earlier candidate should be preferred.
-
-	regions := []RoomToneRegion{
-		{Start: 5 * time.Second, End: 15 * time.Second, Duration: 10 * time.Second},
-		{Start: 30 * time.Second, End: 40 * time.Second, Duration: 10 * time.Second},
-	}
-
-	// Candidate A intervals: ~0.85
-	intervalsA := makeRoomToneTestIntervals(5*time.Second, 10*time.Second,
-		-68.0, -63.0, 100.0, 0.8, 2.0, 0.001)
-	// Candidate B intervals: ~0.86 (slightly higher, within tolerance)
-	intervalsB := makeRoomToneTestIntervals(30*time.Second, 10*time.Second,
-		-69.0, -64.0, 100.0, 0.85, 1.5, 0.0)
-
-	allIntervals := make([]IntervalSample, 0, len(intervalsA)+len(intervalsB))
-	allIntervals = append(allIntervals, intervalsA...)
-	allIntervals = append(allIntervals, intervalsB...)
-
-	result := findBestRoomToneRegion(regions, allIntervals, nil)
-
-	if result.BestRegion == nil {
-		t.Fatal("expected a best region to be selected")
-	}
-
-	// Log scores to verify the gap is within tolerance
-	for _, c := range result.Candidates {
-		t.Logf("candidate start=%v score=%.4f", c.Region.Start, c.Score)
-	}
-
-	if result.BestRegion.Start != 5*time.Second {
-		t.Errorf("BestRegion.Start = %v, want 5s (earlier candidate preferred within tolerance)", result.BestRegion.Start)
-	}
-}
-
-func TestFindBestRoomToneRegion_OnClusterCandidateClearsThresholdDespitePoorSpectral(t *testing.T) {
-	// Under proximity-to-cluster-P50 amplitude scoring, the candidate sitting on the
-	// candidate-population median always earns the full amplitude term (1.0). Two poor
-	// candidates that the former monotonic reward scored below minAcceptableScore (0.3)
-	// now both sit AT the cluster median (identical RMS), so the amplitude term alone
-	// (0.30) clears the threshold and the earliest is elected via the primary branch.
-	//
-	// This documents a deliberate consequence of the amplitude change: with proximity
-	// scoring there is always an on-cluster survivor scoring >= amplitudeScoreWeight, so
-	// the "all surviving candidates below minAcceptableScore" state is no longer reachable
-	// for non-rejected candidates. Structural rejection (crosstalk / digital silence) is
-	// covered separately by TestFindBestRoomToneRegion_AllRejectedCandidatesDoNotFallback.
-
-	regions := []RoomToneRegion{
-		{Start: 0, End: 10 * time.Second, Duration: 10 * time.Second},
-		{Start: 15 * time.Second, End: 25 * time.Second, Duration: 10 * time.Second},
-	}
-
-	// Both candidates identical and poor on spectral/stability: high RMS, voice-range
-	// centroid, flatness 0.0, kurtosis 9.0 (below crosstalk threshold 10), high flux 0.03.
-	// Both sit at clusterP50 (-40.5) so amplitude = 1.0; score = 0.30 + spec + dur + stab.
-	intervalsA := makeRoomToneTestIntervals(0, 10*time.Second,
-		-40.5, -35.5, 2375.0, 0.0, 9.0, 0.03)
-	intervalsB := makeRoomToneTestIntervals(15*time.Second, 10*time.Second,
-		-40.5, -35.5, 2375.0, 0.0, 9.0, 0.03)
-
-	allIntervals := make([]IntervalSample, 0, len(intervalsA)+len(intervalsB))
-	allIntervals = append(allIntervals, intervalsA...)
-	allIntervals = append(allIntervals, intervalsB...)
-
-	result := findBestRoomToneRegion(regions, allIntervals, nil)
-
-	if result.BestRegion == nil {
-		t.Fatal("expected the on-cluster candidate to clear minAcceptableScore and be elected")
-	}
-	if result.BestRegion.Start != 0 {
-		t.Errorf("BestRegion.Start = %v, want 0 (earliest on-cluster candidate within tolerance)", result.BestRegion.Start)
-	}
-	if len(result.Candidates) != 2 {
-		t.Fatalf("len(Candidates) = %d, want 2", len(result.Candidates))
-	}
-	for _, c := range result.Candidates {
-		if c.Score < minAcceptableScore {
-			t.Errorf("candidate start=%v score=%.4f, want >= minAcceptableScore %.1f (amplitude proximity = 1.0)", c.Region.Start, c.Score, minAcceptableScore)
-		}
-	}
-
-	// Log scores for debugging
-	for _, c := range result.Candidates {
-		t.Logf("candidate start=%v score=%.4f", c.Region.Start, c.Score)
-	}
-}
-
-func TestFindBestRoomToneRegion_AllRejectedCandidatesDoNotFallback(t *testing.T) {
-	regions := []RoomToneRegion{
-		{Start: 0, End: 10 * time.Second, Duration: 10 * time.Second},
-		{Start: 15 * time.Second, End: 25 * time.Second, Duration: 10 * time.Second},
-	}
-
-	intervalsA := makeRoomToneTestIntervals(0, 10*time.Second,
-		-120.0, -120.0, 100.0, 0.8, 2.0, 0.0)
-	intervalsB := makeRoomToneTestIntervals(15*time.Second, 10*time.Second,
-		-120.0, -120.0, 100.0, 0.8, 2.0, 0.0)
-
-	allIntervals := make([]IntervalSample, 0, len(intervalsA)+len(intervalsB))
-	allIntervals = append(allIntervals, intervalsA...)
-	allIntervals = append(allIntervals, intervalsB...)
-
-	result := findBestRoomToneRegion(regions, allIntervals, nil)
-
-	if result.BestRegion != nil {
-		t.Fatalf("BestRegion = %+v, want nil for all rejected candidates", result.BestRegion)
-	}
-	if len(result.Candidates) != 2 {
-		t.Fatalf("len(Candidates) = %d, want 2", len(result.Candidates))
-	}
-	for _, c := range result.Candidates {
-		if c.Score != 0.0 {
-			t.Errorf("candidate start=%v score=%.4f, want 0.0", c.Region.Start, c.Score)
-		}
-	}
-}
-
-func TestFindBestRoomToneRegion_NoFallbackBranchNeeded(t *testing.T) {
-	// Documents that the primary election branch always elects when any candidate
-	// survives the structural gates, so the former lower fallback branch was dead.
-	//
-	// clusterP50 is the median RMS of gate-surviving candidates, and medianFloat64
-	// returns an actual member of that set. That member scores calculateAmplitudeScore
-	// = 1.0 (dev = 0), contributing amplitudeScoreWeight (0.30) >= minAcceptableScore
-	// (0.30). So at least one survivor always clears the primary branch's threshold.
-	//
-	// The candidates here have varied RMS around a cluster (median -68), all clean
-	// enough to survive the gates. The earliest-within-tolerance pick is elected by
-	// the primary branch; no fallback is consulted.
-
-	regions := []RoomToneRegion{
-		{Start: 0, End: 10 * time.Second, Duration: 10 * time.Second},
-		{Start: 15 * time.Second, End: 25 * time.Second, Duration: 10 * time.Second},
-		{Start: 30 * time.Second, End: 40 * time.Second, Duration: 10 * time.Second},
-	}
-
-	// Varied RMS around the cluster median (-68). All share strong spectral/stability
-	// metrics so they survive the gates and score comparably (within selectionTolerance),
-	// making the earliest the elected pick.
-	intervalsA := makeRoomToneTestIntervals(0, 10*time.Second,
-		-67.0, -62.0, 100.0, 0.9, 1.0, 0.0)
-	intervalsB := makeRoomToneTestIntervals(15*time.Second, 10*time.Second,
-		-68.0, -63.0, 100.0, 0.9, 1.0, 0.0)
-	intervalsC := makeRoomToneTestIntervals(30*time.Second, 10*time.Second,
-		-69.0, -64.0, 100.0, 0.9, 1.0, 0.0)
-
-	allIntervals := make([]IntervalSample, 0, len(intervalsA)+len(intervalsB)+len(intervalsC))
-	allIntervals = append(allIntervals, intervalsA...)
-	allIntervals = append(allIntervals, intervalsB...)
-	allIntervals = append(allIntervals, intervalsC...)
-
-	result := findBestRoomToneRegion(regions, allIntervals, nil)
-
-	if result.BestRegion == nil {
-		t.Fatal("expected the primary branch to elect a region when candidates survive the gates")
-	}
-	if len(result.Candidates) != 3 {
-		t.Fatalf("len(Candidates) = %d, want 3", len(result.Candidates))
-	}
-
-	maxScore := 0.0
-	for _, c := range result.Candidates {
-		if c.Score > maxScore {
-			maxScore = c.Score
-		}
-		t.Logf("candidate start=%v score=%.4f", c.Region.Start, c.Score)
-	}
-
-	// The elected region must equal the earliest candidate within selectionTolerance
-	// of the maximum score - the same pick the primary loop computes.
-	var wantStart time.Duration
-	for _, c := range result.Candidates {
-		if c.Score >= maxScore-selectionTolerance && c.Score >= minAcceptableScore {
-			wantStart = c.Region.Start
-			break
-		}
-	}
-	if result.BestRegion.Start != wantStart {
-		t.Errorf("BestRegion.Start = %v, want %v (earliest within tolerance, primary branch)", result.BestRegion.Start, wantStart)
-	}
-}
-
-func TestFindBestRoomToneRegion_SingleAcceptableCandidateElected(t *testing.T) {
-	// A single region with an acceptable score should be elected.
-
-	regions := []RoomToneRegion{
-		{Start: 0, End: 10 * time.Second, Duration: 10 * time.Second},
-	}
-
-	// Acceptable score: RMS -65, centroid outside voice range, good flatness
-	intervals := makeRoomToneTestIntervals(0, 10*time.Second,
-		-65.0, -60.0, 100.0, 0.8, 2.0, 0.0)
-
-	result := findBestRoomToneRegion(regions, intervals, nil)
-
-	if result.BestRegion == nil {
-		t.Fatal("expected a best region to be selected")
-	}
-	if result.BestRegion.Start != 0 {
-		t.Errorf("BestRegion.Start = %v, want 0", result.BestRegion.Start)
-	}
-	if len(result.Candidates) != 1 {
-		t.Errorf("len(Candidates) = %d, want 1", len(result.Candidates))
-	}
-
-	t.Logf("single candidate score=%.4f", result.Candidates[0].Score)
-}
-
-func TestFindBestRoomToneRegion_LaterCandidateWinsWhenGapExceedsTolerance(t *testing.T) {
-	// Two candidates where the score gap far exceeds selectionTolerance (0.02).
-	// The later, higher-scoring candidate should win.
-
-	regions := []RoomToneRegion{
-		{Start: 5 * time.Second, End: 15 * time.Second, Duration: 10 * time.Second},
-		{Start: 30 * time.Second, End: 40 * time.Second, Duration: 10 * time.Second},
-	}
-
-	// Candidate A intervals: quiet outlier, 12 dB below the cluster median (amplitude 0)
-	intervalsA := makeRoomToneTestIntervals(5*time.Second, 10*time.Second,
-		-72.0, -67.0, 100.0, 0.8, 2.0, 0.0)
-	// Candidate B intervals: on the cluster median - top amplitude, strong spectral
-	intervalsB := makeRoomToneTestIntervals(30*time.Second, 10*time.Second,
-		-60.0, -55.0, 100.0, 0.9, 1.0, 0.0)
-
-	allIntervals := make([]IntervalSample, 0, len(intervalsA)+len(intervalsB))
-	allIntervals = append(allIntervals, intervalsA...)
-	allIntervals = append(allIntervals, intervalsB...)
-
-	result := findBestRoomToneRegion(regions, allIntervals, nil)
-
-	if result.BestRegion == nil {
-		t.Fatal("expected a best region to be selected")
-	}
-
-	// Log scores to verify the gap exceeds tolerance
-	for _, c := range result.Candidates {
-		t.Logf("candidate start=%v score=%.4f", c.Region.Start, c.Score)
-	}
-
-	if result.BestRegion.Start != 30*time.Second {
-		t.Errorf("BestRegion.Start = %v, want 30s (later candidate wins when gap exceeds tolerance)", result.BestRegion.Start)
-	}
-}
-
-func TestFindBestRoomToneRegion_LateCandidateDiscoverable(t *testing.T) {
-	// A high-scoring candidate placed at 50% of totalDuration (1800s into a 3600s recording)
-	// is correctly elected when it is the only candidate above minAcceptableScore.
-	// Selection evaluates candidates by score regardless of temporal position, with no
-	// excludeFirstSeconds, silenceSearchPercent, or candidateCutoffPercent gate.
-
-	regions := []RoomToneRegion{
-		{Start: 1800 * time.Second, End: 1810 * time.Second, Duration: 10 * time.Second},
-	}
-
-	// High-scoring room tone candidate: low RMS, low centroid, high flatness, low kurtosis, no flux
-	intervals := makeRoomToneTestIntervals(1800*time.Second, 10*time.Second,
-		-75.0, -70.0, 100.0, 0.9, 1.0, 0.0)
-
-	result := findBestRoomToneRegion(regions, intervals, nil)
-
-	if result.BestRegion == nil {
-		t.Fatal("expected candidate at t=1800s to be elected")
-	}
-	if result.BestRegion.Start != 1800*time.Second {
-		t.Errorf("BestRegion.Start = %v, want 1800s", result.BestRegion.Start)
-	}
-	if len(result.Candidates) != 1 {
-		t.Fatalf("len(Candidates) = %d, want 1", len(result.Candidates))
-	}
-
-	t.Logf("late candidate score=%.4f", result.Candidates[0].Score)
-}
-
-// ============================================================================
 // Speech Detection Tests
 // ============================================================================
 
@@ -1000,492 +496,6 @@ func makeSpeechTestIntervals(count int, rms float64) []IntervalSample {
 		}
 	}
 	return intervals
-}
-
-func TestSpeechScore(t *testing.T) {
-	tests := []struct {
-		name         string
-		interval     IntervalSample
-		rmsP50       float64
-		speechRMSMin float64
-		wantMin      float64
-		wantMax      float64
-	}{
-		{
-			name: "typical speech",
-			interval: IntervalSample{
-				RMSLevel: -18.0,
-				Spectral: SpectralMetrics{
-					Centroid: 1500.0, // Voice range
-					Entropy:  0.5,
-				},
-			},
-			rmsP50:       -20.0,
-			speechRMSMin: -40.0,
-			wantMin:      0.4,
-			wantMax:      1.0,
-		},
-		{
-			name: "silence (too quiet)",
-			interval: IntervalSample{
-				RMSLevel: -50.0,
-				Spectral: SpectralMetrics{
-					Centroid: 1500.0,
-					Entropy:  0.5,
-				},
-			},
-			rmsP50:       -20.0,
-			speechRMSMin: -40.0,
-			wantMin:      0.0,
-			wantMax:      0.0,
-		},
-		{
-			name: "noise (wrong centroid)",
-			interval: IntervalSample{
-				RMSLevel: -18.0,
-				Spectral: SpectralMetrics{
-					Centroid: 8000.0, // Outside voice range
-					Entropy:  0.9,    // High entropy
-				},
-			},
-			rmsP50:       -20.0,
-			speechRMSMin: -40.0,
-			wantMin:      0.0,
-			wantMax:      0.4,
-		},
-		{
-			name: "at RMS minimum threshold",
-			interval: IntervalSample{
-				RMSLevel: -40.0, // Exactly at speechRMSMinimum
-				Spectral: SpectralMetrics{
-					Centroid: 1500.0,
-					Entropy:  0.5,
-				},
-			},
-			rmsP50:       -45.0,
-			speechRMSMin: -40.0,
-			wantMin:      0.3,
-			wantMax:      0.85,
-		},
-		{
-			name: "just below RMS minimum",
-			interval: IntervalSample{
-				RMSLevel: -40.1, // Just below threshold
-				Spectral: SpectralMetrics{
-					Centroid: 1500.0,
-					Entropy:  0.5,
-				},
-			},
-			rmsP50:       -45.0,
-			speechRMSMin: -40.0,
-			wantMin:      0.0,
-			wantMax:      0.0,
-		},
-		{
-			name: "low entropy structured speech",
-			interval: IntervalSample{
-				RMSLevel: -18.0,
-				Spectral: SpectralMetrics{
-					Centroid: 2000.0,
-					Entropy:  0.2, // Very structured
-				},
-			},
-			rmsP50:       -20.0,
-			speechRMSMin: -40.0,
-			wantMin:      0.5,
-			wantMax:      1.0,
-		},
-		{
-			name: "adaptive threshold rejects quiet interval",
-			interval: IntervalSample{
-				RMSLevel: -35.0,
-				Spectral: SpectralMetrics{
-					Centroid: 1500.0,
-					Entropy:  0.5,
-				},
-			},
-			rmsP50:       -20.0,
-			speechRMSMin: -32.0, // Adaptive threshold above interval RMS
-			wantMin:      0.0,
-			wantMax:      0.0,
-		},
-		{
-			name: "default threshold admits same interval",
-			interval: IntervalSample{
-				RMSLevel: -35.0,
-				Spectral: SpectralMetrics{
-					Centroid: 1500.0,
-					Entropy:  0.5,
-				},
-			},
-			rmsP50:       -40.0,
-			speechRMSMin: -40.0, // Default threshold below interval RMS
-			wantMin:      0.01,
-			wantMax:      1.0,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			score := speechScore(tt.interval, tt.rmsP50, tt.speechRMSMin)
-			if score < tt.wantMin || score > tt.wantMax {
-				t.Errorf("speechScore() = %.2f, want [%.2f, %.2f]", score, tt.wantMin, tt.wantMax)
-			}
-		})
-	}
-}
-
-func TestComputeSpeechRMSMinimum(t *testing.T) {
-	tests := []struct {
-		name       string
-		rmsLevel   float64
-		noiseFloor float64
-		want       float64
-	}{
-		{
-			name:       "RMS-12 wins over NoiseFloor+6",
-			rmsLevel:   -20.0,
-			noiseFloor: -55.0,
-			want:       -32.0,
-		},
-		{
-			name:       "NoiseFloor+6 wins over RMS-12",
-			rmsLevel:   -20.0,
-			noiseFloor: -35.0,
-			want:       -29.0,
-		},
-		{
-			name:       "fallback on zero measurements",
-			rmsLevel:   0.0,
-			noiseFloor: 0.0,
-			want:       -40.0,
-		},
-		{
-			name:       "fallback on -Inf noise floor",
-			rmsLevel:   -20.0,
-			noiseFloor: math.Inf(-1),
-			want:       -40.0,
-		},
-		{
-			name:       "fallback on -Inf RMS level",
-			rmsLevel:   math.Inf(-1),
-			noiseFloor: -50.0,
-			want:       -40.0,
-		},
-		{
-			name:       "Anna scenario from proposal",
-			rmsLevel:   -48.7,
-			noiseFloor: -70.0,
-			want:       -60.7,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := computeSpeechRMSMinimum(tt.rmsLevel, tt.noiseFloor)
-			if math.Abs(got-tt.want) > 0.001 {
-				t.Errorf("computeSpeechRMSMinimum(%v, %v) = %v, want %v", tt.rmsLevel, tt.noiseFloor, got, tt.want)
-			}
-		})
-	}
-}
-
-func TestFindSpeechCandidatesFromIntervals(t *testing.T) {
-	// Helper to create intervals with varied RMS for realistic median calculation
-	// Speech intervals have majority at/above median to ensure >0.4 speech score
-	// Pattern biased towards louder: 3 intervals at -12 to -16, 1 at -22
-	makeVariedSpeechIntervals := func(startTime time.Duration, count int, isSpeech bool) []IntervalSample {
-		intervals := make([]IntervalSample, count)
-		for i := range intervals {
-			rms := -50.0 // silence
-			if isSpeech {
-				// Create varied RMS where most intervals score above 0.4 threshold
-				// Median will be around -14, most intervals at/above median
-				switch i % 4 {
-				case 0:
-					rms = -12.0 // loud speech (+4 above median -> ampScore ~0.67)
-				case 1:
-					rms = -14.0 // moderate (+2 above median -> ampScore ~0.33)
-				case 2:
-					rms = -16.0 // at median (ampScore ~0)
-				case 3:
-					rms = -22.0 // quiet (below median, ampScore 0)
-				}
-			}
-			intervals[i] = IntervalSample{
-				Timestamp: startTime + time.Duration(i)*250*time.Millisecond,
-				RMSLevel:  rms,
-				Spectral: SpectralMetrics{
-					Centroid: 1500.0,
-					Entropy:  0.3, // Low entropy for better scores (~0.65 entropyScore)
-				},
-			}
-		}
-		return intervals
-	}
-
-	t.Run("finds 30s speech region", func(t *testing.T) {
-		// 10s room tone + 2.5min speech (600 intervals)
-		// With 25% above threshold, this gives ~150 speech intervals (above 120 minimum)
-		// Search starts at 12s, so we lose 8 intervals, leaving 592 → ~148 speech intervals
-		roomToneIntervals := makeVariedSpeechIntervals(0, 40, false)            // 10s room tone
-		speechIntervals := makeVariedSpeechIntervals(10*time.Second, 600, true) // 2.5min speech
-		intervals := make([]IntervalSample, 0, len(roomToneIntervals)+len(speechIntervals))
-		intervals = append(intervals, roomToneIntervals...)
-		intervals = append(intervals, speechIntervals...)
-
-		candidates := findSpeechCandidatesFromIntervals(intervals, roomToneSpan{}, false, -20.0, -55.0)
-
-		if len(candidates) == 0 {
-			t.Fatal("expected at least one speech candidate")
-		}
-		if candidates[0].Duration < 30*time.Second {
-			t.Errorf("duration %v < 30s minimum", candidates[0].Duration)
-		}
-	})
-
-	t.Run("no candidates for short speech", func(t *testing.T) {
-		// 10s room tone + 4s speech (too short - only 16 intervals, below 40 minimum).
-		// findSpeechCandidatesFromIntervals needs >= minimumSpeechIntervals samples,
-		// so pad with room tone after the speech to clear the length guard while
-		// keeping the speech run itself under the 10s gate.
-		roomToneIntervals := makeVariedSpeechIntervals(0, 40, false)
-		speechIntervals := makeVariedSpeechIntervals(10*time.Second, 16, true) // 4s speech
-		tailIntervals := makeVariedSpeechIntervals(14*time.Second, 40, false)  // room-tone tail
-		intervals := make([]IntervalSample, 0, len(roomToneIntervals)+len(speechIntervals)+len(tailIntervals))
-		intervals = append(intervals, roomToneIntervals...)
-		intervals = append(intervals, speechIntervals...)
-		intervals = append(intervals, tailIntervals...)
-
-		candidates := findSpeechCandidatesFromIntervals(intervals, roomToneSpan{}, false, -20.0, -55.0)
-
-		if len(candidates) != 0 {
-			t.Errorf("expected no candidates, got %d", len(candidates))
-		}
-	})
-
-	t.Run("handles natural pauses", func(t *testing.T) {
-		// Speech with 1.5s pause in middle (should bridge)
-		// Each segment needs enough high-scoring intervals: 300 intervals (75s) each
-		speech1 := makeVariedSpeechIntervals(10*time.Second, 300, true)         // 75s
-		pause := makeVariedSpeechIntervals(85*time.Second, 6, false)            // 1.5s pause
-		speech2 := makeVariedSpeechIntervals(86500*time.Millisecond, 300, true) // 75s more
-		intervals := append(append(speech1, pause...), speech2...)
-
-		candidates := findSpeechCandidatesFromIntervals(intervals, roomToneSpan{}, false, -20.0, -55.0)
-
-		if len(candidates) == 0 {
-			t.Fatal("expected speech candidate bridging pause")
-		}
-		// Combined region should be substantial
-		if candidates[0].Duration < 100*time.Second {
-			t.Errorf("expected bridged duration >100s, got %v", candidates[0].Duration)
-		}
-	})
-
-	t.Run("excludes room-tone span without dropping earlier speech", func(t *testing.T) {
-		// Speech before and after an elected room-tone span. Under the decoupled
-		// search both are found, and no candidate spans across the room-tone interval.
-		earlyIntervals := makeVariedSpeechIntervals(0, 200, true)             // 50s speech at start
-		roomTone := makeVariedSpeechIntervals(50*time.Second, 40, false)      // 10s room tone in the gap
-		lateIntervals := makeVariedSpeechIntervals(60*time.Second, 500, true) // 125s speech later
-		intervals := make([]IntervalSample, 0, len(earlyIntervals)+len(roomTone)+len(lateIntervals))
-		intervals = append(intervals, earlyIntervals...)
-		intervals = append(intervals, roomTone...)
-		intervals = append(intervals, lateIntervals...)
-
-		// The elected room tone sits in the 50-60s gap between the two speech runs.
-		span := roomToneSpan{Start: 50 * time.Second, End: 60 * time.Second, Present: true}
-		candidates := findSpeechCandidatesFromIntervals(intervals, span, false, -20.0, -55.0)
-
-		if len(candidates) == 0 {
-			t.Fatal("expected speech candidate before the room-tone span")
-		}
-		// Early speech is now found: the first candidate starts at or near t=0.
-		if candidates[0].Start > 250*time.Millisecond {
-			t.Errorf("early speech start %v should be at or near t=0", candidates[0].Start)
-		}
-		// No candidate spans across the room-tone interval [50s, 60s).
-		for _, c := range candidates {
-			if c.Start < span.End && c.End > span.Start {
-				t.Errorf("candidate [%v, %v) overlaps room-tone span [%v, %v)", c.Start, c.End, span.Start, span.End)
-			}
-		}
-	})
-
-	t.Run("insufficient intervals returns nil", func(t *testing.T) {
-		// Only 30 intervals (7.5s) - less than minimum 40 (10s)
-		intervals := makeVariedSpeechIntervals(0, 30, true)
-
-		candidates := findSpeechCandidatesFromIntervals(intervals, roomToneSpan{}, false, -20.0, -55.0)
-
-		if candidates != nil {
-			t.Errorf("expected nil for insufficient intervals, got %d candidates", len(candidates))
-		}
-	})
-
-	t.Run("long pause breaks speech region", func(t *testing.T) {
-		// Speech with 3s pause in middle (exceeds 2s tolerance, should break).
-		// Each segment is 8s (32 intervals), below the 10s / 40-interval gate, so
-		// neither qualifies alone once the pause splits the run.
-		speech1 := makeVariedSpeechIntervals(10*time.Second, 32, true) // 8s
-		pause := makeVariedSpeechIntervals(18*time.Second, 12, false)  // 3s pause (> tolerance)
-		speech2 := makeVariedSpeechIntervals(21*time.Second, 32, true) // 8s more
-		intervals := append(append(speech1, pause...), speech2...)
-
-		candidates := findSpeechCandidatesFromIntervals(intervals, roomToneSpan{}, false, -20.0, -55.0)
-
-		// Neither segment alone meets the 10s minimum, so no candidates expected
-		if len(candidates) != 0 {
-			t.Errorf("expected no candidates (pause breaks region), got %d", len(candidates))
-		}
-	})
-
-	t.Run("voice-activated bridges digital silence gap", func(t *testing.T) {
-		// Two speech segments separated by a 7.5s digital silence gap (30 intervals).
-		// Default tolerance (8 intervals / 2s) would split; widened (40 / 10s) bridges.
-		speech1 := makeVariedSpeechIntervals(10*time.Second, 300, true)         // 75s speech
-		gap := makeVariedSpeechIntervals(85*time.Second, 30, false)             // 7.5s digital silence
-		speech2 := makeVariedSpeechIntervals(92500*time.Millisecond, 300, true) // 75s speech
-		intervals := append(append(speech1, gap...), speech2...)
-
-		// Default tolerance: gap exceeds 2s, splits into two regions
-		defaultCandidates := findSpeechCandidatesFromIntervals(intervals, roomToneSpan{}, false, -20.0, -55.0)
-		// Widened tolerance: gap within 10s, bridges into one region
-		widenedCandidates := findSpeechCandidatesFromIntervals(intervals, roomToneSpan{}, true, -20.0, -55.0)
-
-		if len(defaultCandidates) < 2 {
-			t.Fatalf("expected default tolerance to split gap into 2+ regions, got %d", len(defaultCandidates))
-		}
-
-		if len(widenedCandidates) != 1 {
-			t.Fatalf("expected widened tolerance to bridge gap into 1 region, got %d", len(widenedCandidates))
-		}
-
-		if widenedCandidates[0].Duration < 140*time.Second {
-			t.Errorf("expected bridged duration >140s, got %v", widenedCandidates[0].Duration)
-		}
-	})
-
-	t.Run("voice-activated does not bridge gaps beyond 10s", func(t *testing.T) {
-		// Two speech segments separated by a 12.5s gap (50 intervals).
-		// Even widened tolerance (40 intervals) should not bridge this.
-		speech1 := makeVariedSpeechIntervals(10*time.Second, 300, true)         // 75s speech
-		gap := makeVariedSpeechIntervals(85*time.Second, 50, false)             // 12.5s gap
-		speech2 := makeVariedSpeechIntervals(97500*time.Millisecond, 300, true) // 75s speech
-		intervals := append(append(speech1, gap...), speech2...)
-
-		candidates := findSpeechCandidatesFromIntervals(intervals, roomToneSpan{}, true, -20.0, -55.0)
-
-		if len(candidates) < 2 {
-			t.Errorf("expected gap >10s to split even with widened tolerance, got %d candidates", len(candidates))
-		}
-	})
-
-	t.Run("speech before a late room tone is found", func(t *testing.T) {
-		// A speech run from t=0, then a late room-tone span, then no speech.
-		// The old floor searched only after the room tone and returned nil; the decoupled
-		// search finds the early speech. The run uses 240 intervals: the varied pattern
-		// scores about a quarter above the median, so ~60 speech intervals clear the 40 gate.
-		speechIntervals := makeVariedSpeechIntervals(0, 240, true)       // 60s speech from t=0
-		roomTone := makeVariedSpeechIntervals(60*time.Second, 40, false) // 10s room tone
-		intervals := make([]IntervalSample, 0, len(speechIntervals)+len(roomTone))
-		intervals = append(intervals, speechIntervals...)
-		intervals = append(intervals, roomTone...)
-
-		span := roomToneSpan{Start: 60 * time.Second, End: 70 * time.Second, Present: true}
-		candidates := findSpeechCandidatesFromIntervals(intervals, span, false, -20.0, -55.0)
-
-		if len(candidates) == 0 {
-			t.Fatal("expected the early speech run to be found")
-		}
-		if candidates[0].Start > 250*time.Millisecond {
-			t.Errorf("candidate start %v should be at or near t=0", candidates[0].Start)
-		}
-		if candidates[0].End > span.Start {
-			t.Errorf("candidate end %v should not extend past the room-tone span start %v", candidates[0].End, span.Start)
-		}
-	})
-
-	t.Run("two runs either side of the span stay distinct", func(t *testing.T) {
-		// A speech run before the span and a separate run after it, with the room-tone
-		// span filled by quiet intervals so masking has samples to mask. The masked
-		// span (40 intervals) is wider than the default 8-interval tolerance, so the
-		// runs must not bridge across the excised span. This is the primary guard
-		// against the splice-vs-mask run-bridging regression.
-		earlyIntervals := makeVariedSpeechIntervals(0, 240, true)             // 60s speech from t=0
-		roomTone := makeVariedSpeechIntervals(60*time.Second, 40, false)      // 10s room tone in the gap
-		lateIntervals := makeVariedSpeechIntervals(70*time.Second, 240, true) // 60s speech later
-		intervals := make([]IntervalSample, 0, len(earlyIntervals)+len(roomTone)+len(lateIntervals))
-		intervals = append(intervals, earlyIntervals...)
-		intervals = append(intervals, roomTone...)
-		intervals = append(intervals, lateIntervals...)
-
-		span := roomToneSpan{Start: 60 * time.Second, End: 70 * time.Second, Present: true}
-		candidates := findSpeechCandidatesFromIntervals(intervals, span, false, -20.0, -55.0)
-
-		if len(candidates) != 2 {
-			t.Fatalf("expected exactly two distinct candidates, got %d", len(candidates))
-		}
-		for _, c := range candidates {
-			if c.Start < span.End && c.End > span.Start {
-				t.Errorf("candidate [%v, %v) spans the room-tone interval [%v, %v)", c.Start, c.End, span.Start, span.End)
-			}
-		}
-	})
-
-	t.Run("no room-tone span searches from t=0", func(t *testing.T) {
-		// The "none" sentinel applies no exclusion; a t=0 speech run is found.
-		intervals := makeVariedSpeechIntervals(0, 240, true) // 60s speech from t=0
-
-		candidates := findSpeechCandidatesFromIntervals(intervals, roomToneSpan{}, false, -20.0, -55.0)
-
-		if len(candidates) == 0 {
-			t.Fatal("expected a candidate with no exclusion applied")
-		}
-		if candidates[0].Start > 250*time.Millisecond {
-			t.Errorf("candidate start %v should be at or near t=0", candidates[0].Start)
-		}
-	})
-
-	t.Run("median computed over non-excluded intervals", func(t *testing.T) {
-		// Loud speech plus a long quiet room-tone span. The quiet intervals outnumber
-		// the speech intervals, so a naive whole-file median would land in the quiet
-		// region and depress every ampScore. Excluding the span from rmsP50 keeps the
-		// median measuring speech-vs-speech, so run membership matches a no-span control.
-		speechIntervals := makeVariedSpeechIntervals(0, 240, true)         // 60s loud speech from t=0
-		quietSpan := makeVariedSpeechIntervals(60*time.Second, 280, false) // 70s quiet room tone
-		intervals := make([]IntervalSample, 0, len(speechIntervals)+len(quietSpan))
-		intervals = append(intervals, speechIntervals...)
-		intervals = append(intervals, quietSpan...)
-
-		span := roomToneSpan{Start: 60 * time.Second, End: 130 * time.Second, Present: true}
-		withSpan := findSpeechCandidatesFromIntervals(intervals, span, false, -20.0, -55.0)
-
-		// Control: same speech, no quiet tail, no span.
-		control := findSpeechCandidatesFromIntervals(speechIntervals, roomToneSpan{}, false, -20.0, -55.0)
-
-		if len(withSpan) == 0 {
-			t.Fatal("expected speech detected with the quiet span excluded from the median")
-		}
-		if len(control) == 0 {
-			t.Fatal("expected speech detected in the no-span control")
-		}
-		if len(withSpan) != len(control) {
-			t.Errorf("run count differs: with span %d, control %d", len(withSpan), len(control))
-		}
-		// Same Start and near-identical duration prove run membership is equivalent.
-		// The small duration slack is the run ending at the span boundary rather than
-		// at EOF; it is the trailing sub-threshold intervals, not a median shift (a
-		// shifted median would collapse the run or move its Start).
-		if withSpan[0].Start != control[0].Start {
-			t.Errorf("run start differs: with span %v, control %v", withSpan[0].Start, control[0].Start)
-		}
-		const boundarySlack = time.Second
-		if d := withSpan[0].Duration - control[0].Duration; d > boundarySlack || d < -boundarySlack {
-			t.Errorf("run duration differs beyond boundary slack: with span %v, control %v", withSpan[0].Duration, control[0].Duration)
-		}
-	})
 }
 
 func TestMeasureSpeechCandidateFromIntervals(t *testing.T) {
@@ -1777,7 +787,7 @@ func TestScoreSpeechIntervalWindow(t *testing.T) {
 							Spectral: SpectralMetrics{
 								Kurtosis: 15.0,    // High
 								Flatness: 0.8,     // Noise-like
-								Centroid: 6000.0,  // Outside voice range
+								Centroid: 7000.0,  // Outside voice range (above speechCentroidMax 6000 Hz)
 								Rolloff:  12000.0, // Above acceptable range (max 10000)
 								Flux:     0.05,    // High variation (transients)
 							},
@@ -1789,7 +799,7 @@ func TestScoreSpeechIntervalWindow(t *testing.T) {
 							Spectral: SpectralMetrics{
 								Kurtosis: 1.0,     // Low
 								Flatness: 0.8,     // Noise-like
-								Centroid: 6000.0,  // Outside voice range
+								Centroid: 7000.0,  // Outside voice range (above speechCentroidMax 6000 Hz)
 								Rolloff:  12000.0, // Above acceptable range (max 10000)
 								Flux:     0.05,    // High variation (transients)
 							},
@@ -1815,8 +825,10 @@ func TestScoreSpeechIntervalWindow(t *testing.T) {
 			name: "low kurtosis (flat spectrum)",
 			setup: func() []IntervalSample {
 				// Low kurtosis (~2), high flatness (~0.8), centroid outside range, quiet
-				// This should score quite low across all metrics
-				return makeSpeechIntervalsScorable(0, 40, 2.0, 0.8, 6000.0, -32.0)
+				// This should score quite low across all metrics.
+				// Centroid 7000 Hz sits above speechCentroidMax (6000 Hz), so it stays
+				// out of the voice range and the centroid term scores 0 as intended.
+				return makeSpeechIntervalsScorable(0, 40, 2.0, 0.8, 7000.0, -32.0)
 			},
 			wantMin: 0.25,
 			wantMax: 0.50,
@@ -2560,48 +1572,6 @@ func TestCalculateStabilityScore(t *testing.T) {
 	}
 }
 
-func TestCrestFactorPenalty(t *testing.T) {
-	tests := []struct {
-		name        string
-		crest       float64
-		peak        float64
-		rms         float64
-		wantPenalty string // "none", "soft", "hard"
-	}{
-		{"clean room tone", 15.0, -60.0, -75.0, "none"},
-		{"elevated crest, safe peak", 35.0, -50.0, -80.0, "soft"},
-		{"danger zone signature", 40.0, -32.0, -74.0, "hard"},
-		{"high crest, clipping peak", 50.0, -10.0, -61.0, "soft"},
-		{"high crest, loud peak outside zone", 60.0, -20.0, -81.0, "soft"},
-		{"boundary: just below soft threshold", 29.9, -35.0, -72.0, "none"},
-		{"boundary: at soft threshold", 30.1, -50.0, -75.0, "soft"},
-		{"boundary: peak at zone low edge", 40.0, -40.0, -74.0, "soft"}, // peak == -40 is NOT > -40
-		{"boundary: peak just inside zone", 40.0, -39.9, -74.0, "hard"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			baseScore := 1.0
-			result := applyCrestFactorPenalty(baseScore, tt.crest, tt.peak, tt.rms)
-
-			var gotPenalty string
-			switch {
-			case result == 1.0:
-				gotPenalty = "none"
-			case result < 0.6: // Hard penalty = 0.5x or less (may compound with soft)
-				gotPenalty = "hard"
-			default:
-				gotPenalty = "soft"
-			}
-
-			if gotPenalty != tt.wantPenalty {
-				t.Errorf("applyCrestFactorPenalty() = %.3f (%s), want %s penalty",
-					result, gotPenalty, tt.wantPenalty)
-			}
-		})
-	}
-}
-
 func TestRunFilterGraph(t *testing.T) {
 	// Generate synthetic audio: 2-second 440Hz tone
 	testFile := generateTestAudio(t, TestAudioOptions{
@@ -2701,390 +1671,6 @@ func TestRunFilterGraphLenientErrors(t *testing.T) {
 		t.Fatal("no frames processed with lenient config")
 	}
 	t.Logf("lenient config processed %d filtered frames", frameCount)
-}
-
-func TestFindBestRoomToneRegion_BoundaryTransientSurvivesAfterRefinement(t *testing.T) {
-	// An 18s candidate where the first 3s contain a boundary transient (speech-to-room-tone
-	// transition) and the remaining 15s are clean room tone. Without pre-scoring refinement,
-	// the spiking transient intervals would trip the RMS-dispersion gate and the candidate
-	// would score 0.0. With refinement, the golden sub-region trims the transient before scoring.
-
-	region := RoomToneRegion{
-		Start:    10 * time.Second,
-		End:      28 * time.Second,
-		Duration: 18 * time.Second,
-	}
-	regions := []RoomToneRegion{region}
-
-	// First 3s: boundary transient - high peak (simulating speech tail), elevated RMS
-	transientIntervals := makeRoomToneTestIntervals(
-		10*time.Second, 3*time.Second,
-		-50.0, -10.0, // RMS -50, peak -10: crest factor 40 dB per interval
-		100.0, 0.8, 2.0, 0.0,
-	)
-
-	// Remaining 15s: clean room tone - low RMS, low peak, good spectral characteristics
-	cleanIntervals := makeRoomToneTestIntervals(
-		13*time.Second, 15*time.Second,
-		-70.0, -65.0, // RMS -70, peak -65: crest factor 5 dB per interval
-		100.0, 0.9, 1.0, 0.0,
-	)
-
-	allIntervals := make([]IntervalSample, 0, len(transientIntervals)+len(cleanIntervals))
-	allIntervals = append(allIntervals, transientIntervals...)
-	allIntervals = append(allIntervals, cleanIntervals...)
-
-	result := findBestRoomToneRegion(regions, allIntervals, nil)
-
-	if result.BestRegion == nil {
-		t.Fatal("expected candidate to be elected after pre-scoring refinement trims boundary transient")
-	}
-
-	// Verify the candidate was refined
-	if len(result.Candidates) == 0 {
-		t.Fatal("expected candidates to be populated")
-	}
-
-	foundRefined := false
-	for _, c := range result.Candidates {
-		if c.WasRefined {
-			foundRefined = true
-
-			if c.OriginalStart != 10*time.Second {
-				t.Errorf("OriginalStart = %v, want 10s", c.OriginalStart)
-			}
-			if c.OriginalDuration != 18*time.Second {
-				t.Errorf("OriginalDuration = %v, want 18s", c.OriginalDuration)
-			}
-
-			// Refined region should exclude the transient at 10-13s
-			if c.Region.Start < 13*time.Second {
-				t.Errorf("Refined region start %v overlaps transient zone (before 13s)", c.Region.Start)
-			}
-
-			// Refinement should trim the transient, leaving a low crest on the clean sub-region
-			if c.CrestFactor > 25.0 {
-				t.Errorf("Refined crest factor = %.1f dB, want < 25.0 dB", c.CrestFactor)
-			}
-
-			// Score should be above minAcceptableScore (0.3) since the clean section is good
-			if c.Score < 0.3 {
-				t.Errorf("Score = %.4f, want >= 0.3 (candidate should pass after refinement)", c.Score)
-			}
-
-			break
-		}
-	}
-
-	if !foundRefined {
-		t.Error("expected WasRefined=true for candidate with boundary transient")
-	}
-}
-
-// makeRoomToneIntervalsFromRMS builds a contiguous interval sequence from explicit
-// per-interval RMS levels, so tests can construct regions with internal RMS dispersion
-// (a spiking interval) or perfectly steady regions. Peak is set 5 dB above each RMS;
-// spectral fields use clean room-tone defaults (low centroid, high flatness, low kurtosis).
-func makeRoomToneIntervalsFromRMS(start time.Duration, rmsLevels []float64) []IntervalSample {
-	intervals := make([]IntervalSample, len(rmsLevels))
-	for i, rms := range rmsLevels {
-		intervals[i] = IntervalSample{
-			Timestamp: start + time.Duration(i)*250*time.Millisecond,
-			RMSLevel:  rms,
-			PeakLevel: rms + 5.0,
-			Spectral: SpectralMetrics{
-				Centroid: 100.0,
-				Flatness: 0.9,
-				Kurtosis: 1.0,
-				Flux:     0.0,
-			},
-		}
-	}
-	return intervals
-}
-
-func TestFindBestRoomToneRegion_ProximityElectsClusterNotQuietOutlier(t *testing.T) {
-	// Regression for the sparse-room-tone defect: a quiet bright outlier (~-79) must not be
-	// elected over a representative cluster (~-74). With proximity-to-cluster-P50
-	// amplitude scoring, the cluster candidates sit on the population median and the
-	// quiet outlier is penalised for departing from it by more than roomToneAmplitudeDecayDB.
-
-	// Regions: two cluster members (~-74) and one quiet bright outlier (~-79).
-	regions := []RoomToneRegion{
-		{Start: 0, End: 10 * time.Second, Duration: 10 * time.Second},
-		{Start: 15 * time.Second, End: 25 * time.Second, Duration: 10 * time.Second},
-		{Start: 30 * time.Second, End: 40 * time.Second, Duration: 10 * time.Second},
-	}
-
-	// Two on-cluster candidates at -74 (steady) define clusterP50 = -74.
-	clusterA := makeRoomToneTestIntervals(0, 10*time.Second,
-		-74.0, -69.0, 100.0, 0.9, 1.0, 0.0)
-	clusterB := makeRoomToneTestIntervals(15*time.Second, 10*time.Second,
-		-74.0, -69.0, 100.0, 0.9, 1.0, 0.0)
-	// Quiet outlier at -79, 5 dB below the cluster (within the 6 dB decay span but
-	// strictly farther from P50 than the on-cluster members, which sit at dev 0).
-	outlier := makeRoomToneTestIntervals(30*time.Second, 10*time.Second,
-		-79.0, -74.0, 100.0, 0.9, 1.0, 0.0)
-
-	allIntervals := make([]IntervalSample, 0, len(clusterA)+len(clusterB)+len(outlier))
-	allIntervals = append(allIntervals, clusterA...)
-	allIntervals = append(allIntervals, clusterB...)
-	allIntervals = append(allIntervals, outlier...)
-
-	result := findBestRoomToneRegion(regions, allIntervals, nil)
-
-	if result.BestRegion == nil {
-		t.Fatal("expected a best region to be selected")
-	}
-	for _, c := range result.Candidates {
-		t.Logf("candidate start=%v rms=%.1f score=%.4f", c.Region.Start, c.RMSLevel, c.Score)
-	}
-
-	// The elected region must be a cluster member (0s or 15s), never the quiet outlier (30s).
-	if result.BestRegion.Start == 30*time.Second {
-		t.Errorf("BestRegion.Start = 30s (quiet outlier elected), want a cluster member at 0s or 15s")
-	}
-
-	// The outlier must score strictly lower than the on-cluster candidates.
-	var clusterScore, outlierScore float64
-	for _, c := range result.Candidates {
-		if c.Region.Start == 0 {
-			clusterScore = c.Score
-		}
-		if c.Region.Start == 30*time.Second {
-			outlierScore = c.Score
-		}
-	}
-	if outlierScore >= clusterScore {
-		t.Errorf("outlier score %.4f >= cluster score %.4f, want outlier strictly lower", outlierScore, clusterScore)
-	}
-}
-
-func TestExceedsRMSDispersion_SpikingIntervalRejected(t *testing.T) {
-	// A region with one interval spiking well beyond 4*MAD from the region RMS median
-	// must trip the dispersion gate. Steady -74 floor with a single -50 spike interval.
-	rmsLevels := []float64{-74, -74, -74, -74, -50, -74, -74, -74, -74, -74}
-	intervals := makeRoomToneIntervalsFromRMS(0, rmsLevels)
-
-	maxDev, exceeds := exceedsRMSDispersion(intervals)
-	if !exceeds {
-		t.Errorf("exceedsRMSDispersion = false (maxDev %.2f MADs), want true for a spiking interval", maxDev)
-	}
-	if maxDev <= roomToneDispersionMADs {
-		t.Errorf("maxDev = %.2f MADs, want > %.1f", maxDev, roomToneDispersionMADs)
-	}
-}
-
-func TestExceedsRMSDispersion_SteadyRegionPasses(t *testing.T) {
-	// A perfectly steady region has MAD ~= 0; the epsilon floor must keep it passing
-	// rather than rejecting any tiny variation. Tests both a flat region and one with
-	// negligible jitter within the epsilon floor.
-	t.Run("perfectly flat", func(t *testing.T) {
-		intervals := makeRoomToneIntervalsFromRMS(0,
-			[]float64{-74, -74, -74, -74, -74, -74, -74, -74})
-		if maxDev, exceeds := exceedsRMSDispersion(intervals); exceeds {
-			t.Errorf("exceedsRMSDispersion = true (maxDev %.2f MADs), want false for a flat region", maxDev)
-		}
-	})
-
-	t.Run("negligible jitter", func(t *testing.T) {
-		// Sub-epsilon dispersion: 0.01 dB jitter, far below roomToneDispersionMADEpsilon.
-		intervals := makeRoomToneIntervalsFromRMS(0,
-			[]float64{-74.00, -74.01, -74.00, -73.99, -74.00, -74.01, -74.00, -73.99})
-		if maxDev, exceeds := exceedsRMSDispersion(intervals); exceeds {
-			t.Errorf("exceedsRMSDispersion = true (maxDev %.2f MADs), want false for negligible jitter", maxDev)
-		}
-	})
-}
-
-func TestFindBestRoomToneRegion_QuietHighCrestRegionAccepted(t *testing.T) {
-	// A representative quiet region with high crest factor (RMS -74, peak -44, crest 30)
-	// was formerly hard-rejected by the fixed silenceCrestFactorMax = 25 dB cap. The
-	// dispersion gate replaces the cap: because every interval shares the same RMS, the
-	// region is steady (MAD ~= 0) and must now be ACCEPTED despite the high crest.
-
-	region := RoomToneRegion{Start: 0, End: 10 * time.Second, Duration: 10 * time.Second}
-	regions := []RoomToneRegion{region}
-
-	// Uniform RMS -74 with peak -44 on every interval: crest 30 dB, but zero RMS dispersion.
-	intervals := makeRoomToneTestIntervals(0, 10*time.Second,
-		-74.0, -44.0, 100.0, 0.9, 1.0, 0.0)
-
-	result := findBestRoomToneRegion(regions, intervals, nil)
-
-	if result.BestRegion == nil {
-		t.Fatal("expected the steady high-crest region to be accepted (dispersion gate, not crest cap)")
-	}
-	if len(result.Candidates) != 1 {
-		t.Fatalf("len(Candidates) = %d, want 1", len(result.Candidates))
-	}
-
-	c := result.Candidates[0]
-	t.Logf("candidate rms=%.1f crest=%.1f score=%.4f warning=%q", c.RMSLevel, c.CrestFactor, c.Score, c.TransientWarning)
-
-	if c.CrestFactor <= 25.0 {
-		t.Errorf("CrestFactor = %.1f dB, want > 25 dB (the region the old cap would reject)", c.CrestFactor)
-	}
-	if c.Score <= 0.0 {
-		t.Errorf("Score = %.4f, want > 0 (steady region must not be rejected)", c.Score)
-	}
-}
-
-// inBandCrestCandidate builds a measured candidate that is in the voice centroid band
-// and below the kurtosis crosstalk threshold, so the population-relative crest-outlier
-// test is the sole crosstalk discriminator. RMS is kept above the digital-silence floor
-// so the candidate enters the crest population.
-func inBandCrestCandidate(rms, crest float64) *RoomToneCandidateMetrics {
-	return &RoomToneCandidateMetrics{RegionSample: RegionSample{RMSLevel: rms, PeakLevel: rms + crest, CrestFactor: crest, Spectral: SpectralMetrics{
-		Centroid: 3000.0, // inside voiceCentroidMin..voiceCentroidMax
-		Kurtosis: 5.0,    // below crosstalkKurtosisThreshold (10)
-		Flatness: 0.5,
-	}}}
-}
-
-func TestIsLikelyCrosstalk_PopulationRelativeCrestOutlier(t *testing.T) {
-	// The fixed 15 dB crest cap was replaced by a population-relative crest-outlier gate:
-	// reject an in-band candidate only when its crest exceeds the candidate-population
-	// crest median by more than roomToneCrosstalkCrestMADs (4) MADs. A quiet broadband
-	// region whose crest is high only because it is quiet - but not an outlier for its
-	// population - must NOT be flagged; a genuine crest outlier must be flagged.
-
-	// Population: a tight cluster of in-band crests around ~30 dB (a typical quiet-room-tone
-	// population where crest is elevated by quietness, not contamination), plus one extreme
-	// outlier. Crest values: 28, 29, 30, 30, 31, 32 and an outlier at 60.
-	measured := []*RoomToneCandidateMetrics{
-		inBandCrestCandidate(-70, 28),
-		inBandCrestCandidate(-70, 29),
-		inBandCrestCandidate(-70, 30),
-		inBandCrestCandidate(-70, 30),
-		inBandCrestCandidate(-70, 31),
-		inBandCrestCandidate(-70, 32),
-		inBandCrestCandidate(-70, 60), // genuine outlier
-	}
-
-	crestMedian, crestMAD := computeCrestOutlierStats(measured)
-	limit := crestMedian + roomToneCrosstalkCrestMADs*crestMAD
-	t.Logf("crestMedian=%.2f crestMAD=%.2f limit=%.2f", crestMedian, crestMAD, limit)
-
-	t.Run("quiet high-crest non-outlier accepted", func(t *testing.T) {
-		// A quiet broadband region with crest 32 dB (above the median, not a population
-		// outlier, well under the old 15 dB cap). Must NOT be flagged crosstalk.
-		c := inBandCrestCandidate(-74, 32)
-		if c.CrestFactor <= crestMedian {
-			t.Fatalf("test setup: crest %.1f should be above median %.1f", c.CrestFactor, crestMedian)
-		}
-		if isLikelyCrosstalk(c, crestMedian, crestMAD, nil) {
-			t.Errorf("isLikelyCrosstalk = true for quiet in-band crest %.1f dB (median %.1f, limit %.1f), want false",
-				c.CrestFactor, crestMedian, limit)
-		}
-	})
-
-	t.Run("genuine crest outlier flagged", func(t *testing.T) {
-		// Crest 60 dB sits far beyond median + 4*MAD. Must be flagged crosstalk.
-		c := inBandCrestCandidate(-70, 60)
-		if c.CrestFactor <= limit {
-			t.Fatalf("test setup: outlier crest %.1f should exceed limit %.1f", c.CrestFactor, limit)
-		}
-		if !isLikelyCrosstalk(c, crestMedian, crestMAD, nil) {
-			t.Errorf("isLikelyCrosstalk = false for crest outlier %.1f dB (median %.1f, limit %.1f), want true",
-				c.CrestFactor, crestMedian, limit)
-		}
-	})
-}
-
-func TestComputeCrestOutlierStats_UniformPopulationEpsilonGuard(t *testing.T) {
-	// A uniform-crest population has MAD ~= 0. Without the epsilon floor, any above-median
-	// candidate would be flagged (median + 4*0 = median). The floor must keep an above-median
-	// candidate passing unless it is a genuine outlier.
-	measured := []*RoomToneCandidateMetrics{
-		inBandCrestCandidate(-70, 30),
-		inBandCrestCandidate(-70, 30),
-		inBandCrestCandidate(-70, 30),
-		inBandCrestCandidate(-70, 30),
-		inBandCrestCandidate(-70, 30),
-	}
-
-	crestMedian, crestMAD := computeCrestOutlierStats(measured)
-	if crestMAD < roomToneCrosstalkCrestMADEpsilon {
-		t.Errorf("crestMAD = %.4f, want >= epsilon %.4f (floor applied)", crestMAD, roomToneCrosstalkCrestMADEpsilon)
-	}
-
-	// A candidate marginally above the uniform median must NOT be flagged: median 30, MAD
-	// floored to 0.05, limit = 30 + 4*0.05 = 30.2. Crest 30.1 is above median but under limit.
-	c := inBandCrestCandidate(-74, 30.1)
-	if isLikelyCrosstalk(c, crestMedian, crestMAD, nil) {
-		t.Errorf("isLikelyCrosstalk = true for crest %.2f in uniform population (median %.2f, MAD %.4f), want false",
-			c.CrestFactor, crestMedian, crestMAD)
-	}
-}
-
-func TestDetectVoiceActivated(t *testing.T) {
-	// Helper to build a candidate slice with a given number of digital silence
-	// and non-digital-silence candidates.
-	makeCandidates := func(digitalSilence, other int, otherWarning string) []RoomToneCandidateMetrics {
-		candidates := make([]RoomToneCandidateMetrics, 0, digitalSilence+other)
-		for range digitalSilence {
-			candidates = append(candidates, RoomToneCandidateMetrics{
-				TransientWarning: "rejected: digital silence (RMS -120.0 dBFS)",
-				Score:            0.0,
-			})
-		}
-		for range other {
-			candidates = append(candidates, RoomToneCandidateMetrics{
-				TransientWarning: otherWarning,
-				Score:            0.0,
-			})
-		}
-		return candidates
-	}
-
-	tests := []struct {
-		name       string
-		candidates []RoomToneCandidateMetrics
-		want       bool
-	}{
-		{
-			name:       "100% digital silence (Anna-style)",
-			candidates: makeCandidates(386, 0, ""),
-			want:       true,
-		},
-		{
-			name:       "99.8% digital silence, 1 crosstalk (Patrick-style)",
-			candidates: makeCandidates(427, 1, "rejected: crosstalk detected"),
-			want:       true,
-		},
-		{
-			name:       "85% digital silence, below threshold (Marius-style)",
-			candidates: makeCandidates(255, 45, ""),
-			want:       false,
-		},
-		{
-			name:       "no candidates",
-			candidates: nil,
-			want:       false,
-		},
-		{
-			name: "candidates with scores pulling ratio down",
-			// 90 digital silence + 10 scored candidates = 90% < 95% threshold
-			candidates: func() []RoomToneCandidateMetrics {
-				cs := makeCandidates(90, 10, "")
-				for i := 90; i < 100; i++ {
-					cs[i].Score = 0.5
-				}
-				return cs
-			}(),
-			want: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := detectVoiceActivated(tt.candidates)
-			if got != tt.want {
-				t.Errorf("detectVoiceActivated() = %v, want %v", got, tt.want)
-			}
-		})
-	}
 }
 
 func TestAnalyseAudio_RoomToneScanDuration(t *testing.T) {
@@ -3304,73 +1890,4 @@ func noiseProfilesEqual(a, b *NoiseProfile) bool {
 		return a == b
 	}
 	return *a == *b
-}
-
-// buildConversationalIntervals shapes intervals like a sparse, conversational
-// speaker: speech runs separated by pauses longer than the
-// 8-interval (2s) tolerance, so no single run reaches the old 120-interval
-// (30s) wall. Each speech run scores above speechScoreThreshold via the same
-// loud/median/quiet RMS pattern the other speech tests use.
-func buildConversationalIntervals(runLen, gapLen, runs int) []IntervalSample {
-	var intervals []IntervalSample
-	t := time.Duration(0)
-	appendBlock := func(count int, isSpeech bool) {
-		for range count {
-			rms := -50.0
-			centroid := 1500.0
-			entropy := 0.6
-			if isSpeech {
-				// Voiced speech: above the gap-dominated median, centred in the
-				// voice band with low entropy so every interval clears
-				// speechScoreThreshold on its own (centroid + entropy alone
-				// reach 0.5, independent of amplitude).
-				rms = -14.0
-				centroid = 2350.0 // voiceMid -> centroidScore 1.0
-				entropy = 0.0     // -> entropyScore 1.0
-			}
-			intervals = append(intervals, IntervalSample{
-				Timestamp: t,
-				RMSLevel:  rms,
-				Spectral: SpectralMetrics{
-					Centroid: centroid,
-					Entropy:  entropy,
-				},
-			})
-			t += 250 * time.Millisecond
-		}
-	}
-	for r := range runs {
-		appendBlock(runLen, true)
-		if r < runs-1 {
-			appendBlock(gapLen, false)
-		}
-	}
-	return intervals
-}
-
-// TestFindSpeechCandidatesFromIntervals_ConversationalSpeakerElects guards the
-// sparse-speech detection fix: lowering minimumSpeechIntervals to 40 (10s) lets sparse,
-// conversational presenters elect a speech candidate. Runs of ~45 speech
-// intervals separated by >2s pauses never sustained 30 unbroken seconds, so the
-// old 120-interval gate yielded zero candidates and the SpeechProfile fell back
-// to whole-file metrics.
-func TestFindSpeechCandidatesFromIntervals_ConversationalSpeakerElects(t *testing.T) {
-	// Three 45-interval speech runs (11.25s each) split by 10-interval (2.5s)
-	// pauses that exceed the 8-interval default tolerance. No run reaches the
-	// old 120 wall, but each clears the new 40-interval gate.
-	const runLen = 45
-	intervals := buildConversationalIntervals(runLen, 10, 3)
-
-	candidates := findSpeechCandidatesFromIntervals(intervals, roomToneSpan{}, false, -20.0, -55.0)
-
-	if len(candidates) == 0 {
-		t.Fatal("expected >=1 speech candidate for conversational speaker at 40-interval gate")
-	}
-
-	// Document the wall: each individual run (45 intervals / 11.25s) is well under
-	// the old 120-interval (30s) threshold, so the pre-fix gate elected nothing.
-	const oldGate = 120
-	if runLen >= oldGate {
-		t.Fatalf("test invariant broken: run length %d should be below the old gate %d", runLen, oldGate)
-	}
 }
